Move common buildscript scripts code into scripts lib
