+++ buildscripts/scripts/get_distros.py
@@ -3,11 +3,16 @@
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
+import sys
 from argparse import ArgumentParser
 from argparse import Namespace as Args
 from collections.abc import Iterable
+from pathlib import Path
 
 import yaml
+
+sys.path.insert(0, Path(__file__).parent.parent.parent.as_posix())
+from buildscripts.scripts.lib.common import flatten, load_editions_file
 
 
 def print_internal_build_artifacts(args: Args, loaded_yaml: dict) -> None:
@@ -50,12 +55,6 @@ def print_distros_for_use_case(args: Args, loaded_yaml: dict) -> None:
 
 def print_editions(args: Args, loaded_yaml: dict) -> None:
     print(" ".join(sorted(loaded_yaml["editions"].keys())))
-
-
-def flatten(list_to_flatten: Iterable[Iterable[str] | str]) -> Iterable[str]:
-    # This is a workaround the fact that yaml cannot "extend" a predefined node which is a list:
-    # https://stackoverflow.com/questions/19502522/extend-an-array-in-yaml
-    return [h for elem in list_to_flatten for h in ([elem] if isinstance(elem, str) else elem)]
 
 
 def test_distro_lists():
@@ -120,8 +119,7 @@ def parse_arguments() -> Args:
 
 def main() -> None:
     args = parse_arguments()
-    with open(args.editions_file) as editions_file:
-        args.func(args, yaml.load(editions_file, Loader=yaml.FullLoader))
+    args.func(args, load_editions_file(args.editions_file))
 
 
 if __name__ == "__main__":
